chore: extract functions so that they can be reused in other tests
